introduced immutable collection interface
