+++ lib/Doctrine/Common/Collections/Collection.php
@@ -19,7 +19,7 @@
 
 namespace Doctrine\Common\Collections;
 
-use Closure, Countable, IteratorAggregate, ArrayAccess;
+use Closure;
 
 /**
  * The missing (SPL) Collection/Array/OrderedMap interface.
@@ -43,7 +43,7 @@ use Closure, Countable, IteratorAggregate, ArrayAccess;
  * @author Jonathan Wage <jonwage@gmail.com>
  * @author Roman Borschel <roman@code-factory.org>
  */
-interface Collection extends Countable, IteratorAggregate, ArrayAccess
+interface Collection extends ImmutableCollection
 {
     /**
      * Adds an element at the end of the collection.
@@ -60,23 +60,6 @@ interface Collection extends Countable, IteratorAggregate, ArrayAccess
      * @return void
      */
     public function clear();
-
-    /**
-     * Checks whether an element is contained in the collection.
-     * This is an O(n) operation, where n is the size of the collection.
-     *
-     * @param mixed $element The element to search for.
-     *
-     * @return boolean TRUE if the collection contains the element, FALSE otherwise.
-     */
-    public function contains($element);
-
-    /**
-     * Checks whether the collection is empty (contains no elements).
-     *
-     * @return boolean TRUE if the collection is empty, FALSE otherwise.
-     */
-    public function isEmpty();
 
     /**
      * Removes the element at the specified index from the collection.
@@ -97,41 +80,6 @@ interface Collection extends Countable, IteratorAggregate, ArrayAccess
     public function removeElement($element);
 
     /**
-     * Checks whether the collection contains an element with the specified key/index.
-     *
-     * @param string|integer $key The key/index to check for.
-     *
-     * @return boolean TRUE if the collection contains an element with the specified key/index,
-     *                 FALSE otherwise.
-     */
-    public function containsKey($key);
-
-    /**
-     * Gets the element at the specified key/index.
-     *
-     * @param string|integer $key The key/index of the element to retrieve.
-     *
-     * @return mixed
-     */
-    public function get($key);
-
-    /**
-     * Gets all keys/indices of the collection.
-     *
-     * @return array The keys/indices of the collection, in the order of the corresponding
-     *               elements in the collection.
-     */
-    public function getKeys();
-
-    /**
-     * Gets all values of the collection.
-     *
-     * @return array The values of all elements in the collection, in the order they
-     *               appear in the collection.
-     */
-    public function getValues();
-
-    /**
      * Sets an element in the collection at the specified key/index.
      *
      * @param string|integer $key   The key/index of the element to set.
@@ -140,121 +88,4 @@ interface Collection extends Countable, IteratorAggregate, ArrayAccess
      * @return void
      */
     public function set($key, $value);
-
-    /**
-     * Gets a native PHP array representation of the collection.
-     *
-     * @return array
-     */
-    public function toArray();
-
-    /**
-     * Sets the internal iterator to the first element in the collection and returns this element.
-     *
-     * @return mixed
-     */
-    public function first();
-
-    /**
-     * Sets the internal iterator to the last element in the collection and returns this element.
-     *
-     * @return mixed
-     */
-    public function last();
-
-    /**
-     * Gets the key/index of the element at the current iterator position.
-     *
-     * @return int|string
-     */
-    public function key();
-
-    /**
-     * Gets the element of the collection at the current iterator position.
-     *
-     * @return mixed
-     */
-    public function current();
-
-    /**
-     * Moves the internal iterator position to the next element and returns this element.
-     *
-     * @return mixed
-     */
-    public function next();
-
-    /**
-     * Tests for the existence of an element that satisfies the given predicate.
-     *
-     * @param Closure $p The predicate.
-     *
-     * @return boolean TRUE if the predicate is TRUE for at least one element, FALSE otherwise.
-     */
-    public function exists(Closure $p);
-
-    /**
-     * Returns all the elements of this collection that satisfy the predicate p.
-     * The order of the elements is preserved.
-     *
-     * @param Closure $p The predicate used for filtering.
-     *
-     * @return Collection A collection with the results of the filter operation.
-     */
-    public function filter(Closure $p);
-
-    /**
-     * Tests whether the given predicate p holds for all elements of this collection.
-     * 
-     * @param Closure $p The predicate.
-     *
-     * @return boolean TRUE, if the predicate yields TRUE for all elements, FALSE otherwise.
-     */
-    public function forAll(Closure $p);
-
-    /**
-     * Applies the given function to each element in the collection and returns
-     * a new collection with the elements returned by the function.
-     *
-     * @param Closure $func
-     *
-     * @return Collection
-     */
-    public function map(Closure $func);
-
-    /**
-     * Partitions this collection in two collections according to a predicate.
-     * Keys are preserved in the resulting collections.
-     *
-     * @param Closure $p The predicate on which to partition.
-     *
-     * @return array An array with two elements. The first element contains the collection
-     *               of elements where the predicate returned TRUE, the second element
-     *               contains the collection of elements where the predicate returned FALSE.
-     */
-    public function partition(Closure $p);
-
-    /**
-     * Gets the index/key of a given element. The comparison of two elements is strict,
-     * that means not only the value but also the type must match.
-     * For objects this means reference equality.
-     *
-     * @param mixed $element The element to search for.
-     *
-     * @return int|string|bool The key/index of the element or FALSE if the element was not found.
-     */
-    public function indexOf($element);
-
-    /**
-     * Extracts a slice of $length elements starting at position $offset from the Collection.
-     *
-     * If $length is null it returns all elements from $offset to the end of the Collection.
-     * Keys have to be preserved by this method. Calling this method will only return the
-     * selected slice and NOT change the elements contained in the collection slice is called on.
-     *
-     * @param int      $offset The offset to start from.
-     * @param int|null $length The maximum number of elements to return, or null for no limit.
-     *
-     * @return array
-     */
-    public function slice($offset, $length = null);
 }
